Rename interface for questioning authority
qa could mean other things (like quality assurance), so spelling it out is less confusing
