+++ src/components/search/QASearchResults.jsx
@@ -6,7 +6,7 @@ import PropTypes from 'prop-types'
 import { clearErrors, appendError } from 'actions/index'
 import { showModal } from 'actions/modals'
 import ResourceTemplateChoiceModal from '../ResourceTemplateChoiceModal'
-import { getTerm } from 'utilities/qa'
+import { getTerm } from 'utilities/QuestioningAuthority'
 import useResource from 'hooks/useResource'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCopy } from '@fortawesome/free-solid-svg-icons'
