+++ src/utilities/QuestioningAuthority.js
